+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const LOGGER_KEY = "logger";
+
+export const OPTIONS_PROVIDER_TOKEN = "pino-options";
